+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `dateTimeCompleted` filter to form store
+
 ## [23.0.0] - 2025-07-01
 
 ### Added