+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [23.0.1] - 2025-07-04
+
 ### Changed
 
 - `@oneblink/storage`
+
+### Dependencies
+
+- update [@oneblink/storage](https://www.npmjs.com/package/@oneblink/storage) to [3.1.1-beta.1](https://github.com/oneblink/storage/blob/master/CHANGELOG.md) (from [3.1.0-beta.1](https://github.com/oneblink/storage/blob/master/CHANGELOG.md))
 
 ## [23.0.0] - 2025-07-01